+++ P-Emerald_2/arm9/source/keyboard.cpp
@@ -44,10 +44,10 @@ extern SpriteInfo spriteInfo[ SPRITE_COUNT ];
 extern SpriteInfo spriteInfoTop[ SPRITE_COUNT ];
 extern OAMTable *Oam, *OamTop;
 
-std::wstring keyboard::getText( int p_length, const char* p_msg, bool p_time ) {
+std::wstring keyboard::getText( u16 p_length, const char* p_msg, bool p_time ) {
     drawSub( );
 
-    for( int i = 0; i < 4; ++i )
+    for( u8 i = 0; i < 4; ++i )
         Oam->oamBuffer[ 31 + 2 * i ].isHidden = true;
     Oam->oamBuffer[ A_ID ].isHidden = false;
     Oam->oamBuffer[ BWD_ID ].isHidden = false;
@@ -82,7 +82,7 @@ std::wstring keyboard::getText( int p_length, const char* p_msg, bool p_time ) {
     std::wstring out;
     consoleSelect( &Bottom );
     touchPosition t;
-    int i = 0;
+    u8 i = 0;
     while( 1 ) {
         scanKeys( );
         touchRead( &t );
@@ -171,8 +171,8 @@ void keyboard::draw( ) {
     switch( _ind ) {
         case 0://UPPER
         {
-            int cnt = 0;
-            int rw_cnt = 9, cl_cnt = 3;
+            u8 cnt = 0;
+            u8 rw_cnt = 9, cl_cnt = 3;
             _chars.clear( );
             for( char i = 'A'; i <= 'Z'; i++ ) {
                 if( cnt++ == 13 ) {
@@ -188,7 +188,7 @@ void keyboard::draw( ) {
             rw_cnt += 4;
             cl_cnt = 3;
             char pt[ 14 ] = "ÄÖÜ .!?()[]{}";
-            for( int i = 0; i < 13; ++i ) {
+            for( u8 i = 0; i < 13; ++i ) {
                 cust_font2.printChar( pt[ i ], cl_cnt * 8, rw_cnt * 8, true );
                 _chars[ std::pair<int, int>( cl_cnt, rw_cnt ) ] = pt[ i ];
                 _chars[ std::pair<int, int>( cl_cnt, rw_cnt + 1 ) ] = pt[ i ];
@@ -198,8 +198,8 @@ void keyboard::draw( ) {
         }
         case 1://LOWER
         {
-            int cnt = 0;
-            int rw_cnt = 9, cl_cnt = 3;
+            u8 cnt = 0;
+            u8 rw_cnt = 9, cl_cnt = 3;
             for( char i = 'a'; i <= 'z'; i++ ) {
                 if( cnt++ == 13 ) {
                     cnt = 0;
@@ -214,7 +214,7 @@ void keyboard::draw( ) {
             rw_cnt += 4;
             cl_cnt = 3;
             char pt[ 14 ] = "äöüßé §$%&/\\#";
-            for( int i = 0; i < 13; ++i ) {
+            for( u8 i = 0; i < 13; ++i ) {
                 cust_font2.printChar( pt[ i ], cl_cnt * 8, rw_cnt * 8, true );
                 _chars[ std::pair<int, int>( cl_cnt, rw_cnt ) ] = pt[ i ];
                 _chars[ std::pair<int, int>( cl_cnt, rw_cnt + 1 ) ] = pt[ i ];
@@ -224,10 +224,10 @@ void keyboard::draw( ) {
         }
         case 2://NUM
         {
-            int rw_cnt = 9, cl_cnt = 3;
+            u8 rw_cnt = 9, cl_cnt = 3;
 
             char pt[ 14 ] = "0123456789,;:";
-            for( int i = 0; i < 13; ++i ) {
+            for( u8 i = 0; i < 13; ++i ) {
                 cust_font2.printChar( pt[ i ], cl_cnt * 8, rw_cnt * 8, true );
                 _chars[ std::pair<int, int>( cl_cnt, rw_cnt ) ] = pt[ i ];
                 _chars[ std::pair<int, int>( cl_cnt, rw_cnt + 1 ) ] = pt[ i ];
@@ -235,7 +235,7 @@ void keyboard::draw( ) {
             }
             rw_cnt += 3;
             cl_cnt = 3;
-            for( int i = 3 * 49 - 10; i < 3 * 49 + 3; ++i ) {
+            for( u8 i = 3 * 49 - 10; i < 3 * 49 + 3; ++i ) {
                 cust_font2.printChar( i, cl_cnt * 8, rw_cnt * 8, true );
                 _chars[ std::pair<int, int>( cl_cnt, rw_cnt ) ] = i;
                 _chars[ std::pair<int, int>( cl_cnt, rw_cnt + 1 ) ] = i;
@@ -243,7 +243,7 @@ void keyboard::draw( ) {
             }
             rw_cnt += 4;
             cl_cnt = 3;
-            for( int i = 3 * 49 + 3; i < 3 * 49 + 16; ++i ) {
+            for( u8 i = 3 * 49 + 3; i < 3 * 49 + 16; ++i ) {
                 cust_font2.printChar( i, cl_cnt * 8, rw_cnt * 8, true );
                 _chars[ std::pair<int, int>( cl_cnt, rw_cnt ) ] = i;
                 _chars[ std::pair<int, int>( cl_cnt, rw_cnt + 1 ) ] = i;
